# doc/openApi.yaml
openapi: 3.0.3
info:
  title: Music Favorites API
  version: 1.0.0
  description: REST API for managing users, artists, albums, tracks, and favorites.
paths:
...
  /track:
    get:
      summary: Get all tracks
      responses:
        '200':
          description: All tracks
    post:
      summary: Create track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
      responses:
        '201':
          description: Track created
        '400':
          description: Missing fields
  /track/{id}:
    get:
      summary: Get track by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Track found
        '400':
          description: Invalid UUID
        '404':
          description: Not found
    put:
      summary: Update track
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Track'
      responses:
        '200':
          description: Track updated
        '400':
          description: Invalid UUID
        '404':
          description: Not found
    delete:
      summary: Delete track
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Track deleted
        '400':
          description: Invalid UUID
        '404':
          description: Not found
...
  /favs/track/{id}:
    post:
      summary: Add track to favorites
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '201':
          description: Added
        '400':
          description: Invalid UUID
        '422':
          description: Track does not exist
    delete:
      summary: Remove track from favorites
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Removed
        '400':
          description: Invalid UUID
        '404':
          description: Not in favorites
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the resource
  schemas:
...
    Track:
      type: object
      required: [name, duration]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        duration:
          type: number
        artistId:
          type: string
          format: uuid
          nullable: true
        albumId:
          type: string
          format: uuid
          nullable: true